+++ commands/fun/cat.js
@@ -9,14 +9,17 @@ module.exports = {
 
     const Discord = require("discord.js");
     if (!message.guild) return;
-    const superagent = require("superagent");
-
-    let { body } = await superagent.get(`http://api.thecatapi.com/v1/images/search`);
+    const fetch = require("node-fetch");
+  
+    let url;
+    await fetch(`https://nekos.life/api/v2/img/meow`)
+              .then(res => res.json())
+              .then(json => url = json.url)
 
     const catEmbed = new Discord.MessageEmbed()
       .setTitle("Aww... Kitty!")
       .setColor("#ff9900")
-      .setImage(body[0].url)
+      .setImage(url)
       .setFooter(`${message.guild.name}`, `https://i.imgur.com/BkZY6H8.png`);
 
     message.channel.send(catEmbed);